+++ src/main.tsx
@@ -1,3 +1,5 @@
 import '@beacio/core/auto';
+import { installPwaUpdateReload } from './app/pwaUpdate';
 
+installPwaUpdateReload();
 void import('./bootstrap');